requirements.txt
django==2.2
djangorestframework
requests
djoser
requests_oauthlib
lxml

# linting
pylint
pylint-django